requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
motor>=3.4.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1
nltk>=3.8.1
google-genai>=0.3.0
python-dotenv>=1.2.2
pytest>=8.1.1
httpx>=0.27.0
pydantic[email]>=2.7.0
python-multipart>=0.0.9
pydantic-settings>=2.2.1
dnspython>=2.6.1
certifi>=2024.2.2
feedparser>=6.0.10
groq>=0.5.0
